opcua-server-sim/requirements.txt
asyncua==1.0.8
cryptography==41.0.7
PyYAML==6.0.1
